Fix log in PROPPATCH. Relabel {Owned,}Name to {Owned,}XmlName. Move WebDAV-only utils to webdav module. Port rmpty/allprop/propname XML Names from webdav-handler-rs
