+++ Cargo.toml
@@ -4,7 +4,7 @@ name = "https"
 description = "Host These Things Please - a basic http server for hosting a folder fast and simply"
 repository = "https://github.com/thecoshman/http"
 readme = "README.md"
-keywords = ["http", "server", "https", "file", "directory"]
+keywords = ["http", "server", "https", "webdav", "directory"]
 categories = ["network-programming", "web-programming::http-server"]
 license = "MIT"
 build = "build.rs"
